Unify format of REST error responses
Modify loopback.errorHandler() to call the error handler provided by
strong-remoting's RestAdapter when the client accepts json, and fall
back to connect/express handler otherwise.

This way when a LoopBack client requests a URL not mapped to any shared
method, the server returns response in the same format as is produced
in other cases.