Require data-sources directory in addition to models directory.
This prevents a potential issue where a model is loaded from config that specifies a datasource
that is loaded via JS. The model would typically be loaded before the datasource in this case,
leading to an error.

This allows the files in `data-sources` to be loaded with the full `appConfig` attached
to the app object.

Signed-off-by: Samuel Reed <sam@tixelated.com>