first pass at using email-templates module
I haven’t done the work to ensure backwards compatibility with the old
template system and the options that were passed.  Will work on some
tests for that soon.

The lazy load might be unnecessary but I didn’t find a way to make
async calls at this stage of initialization.