Fix emailVerified becoming false on user model update
Yesterday, the loopback we are using in our system was upgraded via npm, and since the upgrade, we noticed that every time the user model updates, the emailVerified column would change to false.

I took a look and realized there might be an error in https://github.com/strongloop/loopback/commit/eb640d8da0cadfbb88171bb823a3af0f60ef5cd9

The intent of the commit just mention is to make emailVerified false when the email gets changed, but notice that ctx.data.email is null on updates, so the condition is always met and emailVerified always becomes false.

This commit fixes the issue just mentioned.