Fix context propagation broken by async@2.x
Rework the REST middleware to use a hand-written version of
"async.eachSeries". Before this change, we were loosing CLS context
when the application was relying on the REST middleware to load the
context middleware.

This is fixing a problem introduced by post-1.0 versions of async,
which we upgraded to via fea3b781a.