Support middleware injected by OpenTelemetry
OpenTelemetry injects a proxy object into the router stack, which it
uses for its tracing.  This is similar to how NewRelic
adds a sentinel handler to the router stack. This commit adds a
similar workaround so that loopback can find the original layer.

Made it more generic to just test if shimmer is used to wrap the handler